+++ quadraticfit.h
@@ -4,12 +4,12 @@
 #include "utils.h"
 
 template <size_t N>
-float Evaluate(const std::array<float, N * 2 + 1>& coefficients, const std::vector<float>& row, const std::array<int, N>& columnIndices)
+double Evaluate(const std::array<double, N * 2 + 1>& coefficients, const std::vector<double>& row, const std::array<int, N>& columnIndices)
 {
-    float ret = coefficients[N * 2];
+    double ret = coefficients[N * 2];
     for (size_t i = 0; i < N; ++i)
     {
-        float x = row[columnIndices[i]];
+        double x = row[columnIndices[i]];
         ret += coefficients[i * 2 + 0] * x * x;
         ret += coefficients[i * 2 + 1] * x;
     }
@@ -17,19 +17,19 @@ float Evaluate(const std::array<float, N * 2 + 1>& coefficients, const std::vect
 }
 
 template <size_t N>
-float RSquared(const std::array<float, N * 2 + 1>& coefficients, const CSV& data, const std::array<int, N>& columnIndices, int valueIndex)
+double RSquared(const std::array<double, N * 2 + 1>& coefficients, const CSV& data, const std::array<int, N>& columnIndices, int valueIndex)
 {
     Average averageSales;
     for (const auto& row : data.data)
         averageSales.AddSample(row[valueIndex]);
 
-    float numerator = 0.0f;
-    float denominator = 0.0f;
+    double numerator = 0.0f;
+    double denominator = 0.0f;
     for (const auto& row : data.data)
     {
-        float actual = row[valueIndex];
+        double actual = row[valueIndex];
 
-        float estimate = Evaluate(coefficients, row, columnIndices);
+        double estimate = Evaluate(coefficients, row, columnIndices);
 
         numerator += sqr(actual - estimate);
         denominator += sqr(actual - averageSales.average);
@@ -39,31 +39,31 @@ float RSquared(const std::array<float, N * 2 + 1>& coefficients, const CSV& data
 }
 
 template <size_t N>
-float AdjustedRSquared(const std::array<float, N * 2 + 1>& coefficients, const CSV& data, const std::array<int, N>& columnIndices, int valueIndex)
+double AdjustedRSquared(const std::array<double, N * 2 + 1>& coefficients, const CSV& data, const std::array<int, N>& columnIndices, int valueIndex)
 {
     int predictorCount = int(N);
     int numSamples = (int)data.data.size();
 
-    float rsquared = RSquared(coefficients, data, columnIndices, valueIndex);
+    double rsquared = RSquared(coefficients, data, columnIndices, valueIndex);
     
-    float numerator = (1.0f - rsquared) * float(numSamples - 1);
-    float denominator = float(numSamples - predictorCount - 1);
+    double numerator = (1.0f - rsquared) * double(numSamples - 1);
+    double denominator = double(numSamples - predictorCount - 1);
 
     return 1.0f - numerator / denominator;
 }
 
 template <size_t N>
-float LossFunction(const std::array<float, N * 2 + 1>& coefficients, const CSV& data, const std::array<int, N>& columnIndices, int valueIndex)
+double LossFunction(const std::array<double, N * 2 + 1>& coefficients, const CSV& data, const std::array<int, N>& columnIndices, int valueIndex)
 {
     Average MSE;
 
     for (const auto& row : data.data)
     {
-        float estimate = Evaluate(coefficients, row, columnIndices);
+        double estimate = Evaluate(coefficients, row, columnIndices);
 
-        float actual = row[valueIndex];
+        double actual = row[valueIndex];
 
-        float error = estimate - actual;
+        double error = estimate - actual;
 
         MSE.AddSample(error * error);
     }
@@ -72,18 +72,18 @@ float LossFunction(const std::array<float, N * 2 + 1>& coefficients, const CSV& 
 }
 
 template <size_t N>
-void CalculateGradient(std::array<float, N * 2 + 1>& gradient, const std::array<float, N * 2 + 1>& _coefficients, const CSV& data, const std::array<int, N>& columnIndices, int valueIndex)
+void CalculateGradient(std::array<double, N * 2 + 1>& gradient, const std::array<double, N * 2 + 1>& _coefficients, const CSV& data, const std::array<int, N>& columnIndices, int valueIndex)
 {
     // Calculates a gradient via central differences
     for (size_t index = 0; index <= N * 2; ++index)
     {
-        std::array<float, N * 2 + 1> coefficients = _coefficients;
+        std::array<double, N * 2 + 1> coefficients = _coefficients;
 
         coefficients[index] = _coefficients[index] - c_epsilon;
-        float A = LossFunction(coefficients, data, columnIndices, valueIndex);
+        double A = LossFunction(coefficients, data, columnIndices, valueIndex);
 
         coefficients[index] = _coefficients[index] + c_epsilon;
-        float B = LossFunction(coefficients, data, columnIndices, valueIndex);
+        double B = LossFunction(coefficients, data, columnIndices, valueIndex);
 
         gradient[index] = (B - A) / (2.0f * c_epsilon);
     }
